test: port the fan-out guard tests from ReactiveResponse onto compose()
Refs #749
